Add `release_on_start` task configuration
Passing `release_on_start=True` to a singleton task makes the task
release the lock when the job starts running, rather than when it has
finished.

Note: This is only available in celery versions 5.2 and above since it
uses the `before_start`-method on the celery Task object which was
introduced in version 5.2.